+++ identity/client.go
@@ -151,8 +151,8 @@ func (c *Client) PubliclyRevealAttributes(
 		c.Options.ProtocolID,
 		c.Options.KeyID,
 		wallet.Counterparty{Type: wallet.CounterpartyTypeAnyone},
-		true,  // forSelf
-		true,  // includeSignature
+		true, // forSelf
+		true, // includeSignature
 		pushdrop.LockBefore,
 	)
 	if err != nil {